#448: Moved code to util
