+++ src/scss/_panktee.scss
@@ -4,7 +4,7 @@
 
     &.larivaar {
       // word-break: break-all;
-    }  
+    }
   }
 
   .wordBreakSpecialChar {
